pingcrm/requirements.txt
Django==3.0.2
djangorestframework==3.11.0
psycopg2-binary==2.8.3
django-environ==0.4.5
ipython[all]
django-webpack-loader
